refactor: centralize resolveDataRoot utility
