+++ apps/cms/src/app/api/import-products/[shop]/route.ts
@@ -1,20 +1,9 @@
 import { authOptions } from "@cms/auth/options";
 import { getServerSession } from "next-auth";
 import { NextResponse, type NextRequest } from "next/server";
-import fsSync, { promises as fs } from "node:fs";
+import { promises as fs } from "node:fs";
 import path from "node:path";
-
-function resolveDataRoot(): string {
-  let dir = process.cwd();
-  while (true) {
-    const candidate = path.join(dir, "data", "shops");
-    if (fsSync.existsSync(candidate)) return candidate;
-    const parent = path.dirname(dir);
-    if (parent === dir) break;
-    dir = parent;
-  }
-  return path.resolve(process.cwd(), "data", "shops");
-}
+import { resolveDataRoot } from "@platform-core/utils";
 
 export async function POST(
   req: NextRequest,
