+++ packages/lib/src/checkShopExists.server.ts
@@ -1,22 +1,10 @@
 // packages/lib/checkShopExists.server.ts
 import "server-only";
 
-import * as fsSync from "node:fs";
 import { promises as fs } from "node:fs";
 import * as path from "node:path";
+import { resolveDataRoot } from "@platform-core/utils";
 import { validateShopName } from "./validateShopName";
-
-function resolveDataRoot(): string {
-  let dir = process.cwd();
-  while (true) {
-    const candidate = path.join(dir, "data", "shops");
-    if (fsSync.existsSync(candidate)) return candidate;
-    const parent = path.dirname(dir);
-    if (parent === dir) break;
-    dir = parent;
-  }
-  return path.resolve(process.cwd(), "data", "shops");
-}
 
 const DATA_ROOT = resolveDataRoot();
