+++ packages/platform-core/src/utils/index.ts
@@ -1,0 +1,3 @@
+export { getShopFromPath } from "./getShopFromPath";
+export { replaceShopInPath } from "./replaceShopInPath";
+export { resolveDataRoot } from "./resolveDataRoot";
